+++ src/main/resources/db/migration/V1__Create_match_table.sql
@@ -4,5 +4,5 @@ create table matcheam.match (
     date DATETIME　not null,
     gameTime TIME not null,
     level varchar(100) not null,
-    maxPlayers int
+    maxPlayers DECIMAL(100)
 );
